requirements.txt
discord.py
aiohttp
aiofiles
unidecode
numpy
scikit-learn
plotly
pandas
dropbox
openai
kaleido
